Поправляет код в задании 7. Личный проект: шаблонизируй это
