+++ .eslintrc.yml
@@ -6,4 +6,7 @@ env:
   es6: true
   browser: true
 
+rules:
+  prefer-const: error
+
 extends: 'htmlacademy/es6'
