+++ src/view/trip-point.js
@@ -1,5 +1,8 @@
-export const createTripPointTemplate = (wayPoint) => {
-  const {dayNumber, currentDate} = wayPoint;
+export const createTripPointTemplate = (wayPoint, index) => {
+  const {currentDate} = wayPoint;
+
+  const DAY_GAP = 1;
+  const dayNumber = index + DAY_GAP;
 
   const dateByDataTime = (date) => {
     return `${date.toLocaleString(`en-US`, {year: `numeric`})}-${date.toLocaleString(`en-US`, {month: `2-digit`})}-${date.toLocaleString(`en-US`, {day: `2-digit`})}`;
